locate img logic and display
